LPS-170922 Moves classes to site-api module
