+++ modules/apps/portal-security/portal-security-membership-policy-test/src/testIntegration/java/com/liferay/portal/security/membership/policy/organization/test/OrganizationMembershipPolicyRowCheckerTest.java
@@ -27,8 +27,8 @@ import com.liferay.portal.security.membership.policy.organization.BaseOrganizati
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 import com.liferay.portlet.RenderResponseFactory;
-import com.liferay.portlet.sites.search.OrganizationRoleUserChecker;
 import com.liferay.portlet.usersadmin.search.UserOrganizationChecker;
+import com.liferay.site.search.OrganizationRoleUserChecker;
 
 import javax.portlet.RenderResponse;
 import javax.portlet.filter.RenderResponseWrapper;
