+++ modules/apps/site/site-api/src/main/java/com/liferay/site/search/UserGroupRoleRoleChecker.java
@@ -12,42 +12,41 @@
  * details.
  */
 
-package com.liferay.portlet.sites.search;
+package com.liferay.site.search;
 
-import com.liferay.portal.kernel.dao.search.RowChecker;
+import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Organization;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.membershippolicy.OrganizationMembershipPolicyUtil;
+import com.liferay.portal.kernel.security.membershippolicy.SiteMembershipPolicyUtil;
 import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.service.UserGroupRoleLocalServiceUtil;
 
 import javax.portlet.RenderResponse;
 
 /**
- * @author Roberto Díaz
+ * @author Jorge Ferrer
  */
-public class OrganizationRoleUserChecker extends RowChecker {
+public class UserGroupRoleRoleChecker extends EmptyOnClickRowChecker {
 
-	public OrganizationRoleUserChecker(
-		RenderResponse renderResponse, Organization organization, Role role) {
+	public UserGroupRoleRoleChecker(
+		RenderResponse renderResponse, User user, Group group) {
 
 		super(renderResponse);
 
-		_organization = organization;
-		_role = role;
+		_user = user;
+		_group = group;
 	}
 
 	@Override
 	public boolean isChecked(Object object) {
-		User user = (User)object;
+		Role role = (Role)object;
 
 		try {
 			return UserGroupRoleLocalServiceUtil.hasUserGroupRole(
-				user.getUserId(), _organization.getGroupId(),
-				_role.getRoleId());
+				_user.getUserId(), _group.getGroupId(), role.getRoleId());
 		}
 		catch (Exception exception) {
 			_log.error(exception);
@@ -58,25 +57,25 @@ public class OrganizationRoleUserChecker extends RowChecker {
 
 	@Override
 	public boolean isDisabled(Object object) {
-		User user = (User)object;
+		Role role = (Role)object;
 
 		try {
-			if (isChecked(user)) {
-				if (OrganizationMembershipPolicyUtil.isRoleProtected(
+			if (isChecked(role)) {
+				if (SiteMembershipPolicyUtil.isRoleProtected(
 						PermissionThreadLocal.getPermissionChecker(),
-						user.getUserId(), _organization.getOrganizationId(),
-						_role.getRoleId()) ||
-					OrganizationMembershipPolicyUtil.isRoleRequired(
-						user.getUserId(), _organization.getOrganizationId(),
-						_role.getRoleId())) {
+						_user.getUserId(), _group.getGroupId(),
+						role.getRoleId()) ||
+					SiteMembershipPolicyUtil.isRoleRequired(
+						_user.getUserId(), _group.getGroupId(),
+						role.getRoleId())) {
 
 					return true;
 				}
 			}
 			else {
-				if (!OrganizationMembershipPolicyUtil.isRoleAllowed(
-						user.getUserId(), _organization.getOrganizationId(),
-						_role.getRoleId())) {
+				if (!SiteMembershipPolicyUtil.isRoleAllowed(
+						_user.getUserId(), _group.getGroupId(),
+						role.getRoleId())) {
 
 					return true;
 				}
@@ -90,9 +89,9 @@ public class OrganizationRoleUserChecker extends RowChecker {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		OrganizationRoleUserChecker.class);
+		UserGroupRoleRoleChecker.class);
 
-	private final Organization _organization;
-	private final Role _role;
+	private final Group _group;
+	private final User _user;
 
 }
